[No JIRA] Fix duplication error with icon font
For `Backpack/Icon` we need to use a font with a different name than the
icon font used in React Native, otherwise UIKit will refuse to load both
of these at the same time. This has been solved by using a new
duplicated font in `bpk-svgs` with a different font name internally.