[No JIRA] Improve `BPKDialogView` in tiny viewport
`BPKDialogView` and its parent `BPKDialogController` did not handle
extreme cases with small viewports well. On iPhone SE the dialog would
overflow the screen and make it impossible to interact with the relevant
actions.

In this commit this has been improved in a few ways:

1. The dialog will now shrink the text size of the message.
2. The dialog will eventually truncate the message if needed.
3. The dialog's title is limited to two lines.
4. In `BPKDialogController` the dialog heigh is limited to fit in the
viewport vertically as well as horizontally.

As part of this I refactored the common parts of the component and
resolved several issue with ambiguous height and vertical position.