[BPK-4208] Make CI work with Xcode 12
As part of this commit the following changes have been made:

+ Disabled arm64 for simulator builds, this is presumably something that
has changed with Xcode 12 and Apple Silicon to prepare the simulator to
run on arm64 in the future.
+ Use Xcode 12 and iOS 13.5 for the secondary build. Having the build on
Xcode 11 proved to be impossible while also having the project
compatible with Xcode 12. Travis' Xcode 12 image does not include iOS
13.7 at the time of writing but we should upgrade to it in the future.
+ Set `ONLY_ACTIVE_ARCH=YES` for CI builds, again without this the
builds for arm64 would fail.
+ Tweak toats UI tests, these were failing due to timing issues.