[BPK-3625] Migrate to GitHub Actions for CI
Set up GitHub action to mirror Travis in preparation for the eventual
removal of Travis. Danger is currently disabled as we need to work out a
few issues with it(possibly migrate to Danger.js which is better
maintained).

I definitely did this in one commit not 30 + a rebase, you can trust me
😇