[KOA-4294] Move BPKHorizontalNavigation to Swift
Migrate `BPKHorizontalNavigation` to Swift instead of Objective-C to
support an API that enforces the appropriate use of icons at compile
time rather than runtime.

A shim is provided for Objective-C that supports part of the API, but
not all. Consumers using Objective-C are advised to migrate to Swift or
create wrapping views in Swift that are Objective-C compatible.