[DO NOT MERGE] Add peeking bottom sheet functionality (#2017)
* feat: Add peeking bottom sheet functionality

The code changes introduce a new feature to add peeking bottom sheet functionality. This allows the user to show a partially visible bottom sheet that can be expanded by dragging. The `BottomSheetExampleView` now includes a button to show the peeking bottom sheet. The `ContentFitBottomSheet` struct has been updated to include a `peekHeight` parameter, allowing the user to specify the initial height of the peeking bottom sheet.

* feat: Add peeking bottom sheet functionality

* feat: Add peeking bottom sheet functionality

* feat: Update presentation background interaction in BPKBottomSheet.swift

The code changes update the presentation background interaction in BPKBottomSheet.swift to use the `PresentationBackgroundInteraction.enabled` enum value instead of the deprecated `.enabled` shorthand. This ensures compatibility with the latest SwiftUI version and improves the overall functionality of the bottom sheet component.